Refactor tests into smaller pieces
